Let institutional admins see all last institutions
Before institutional admins that weren't site admins went to the members
page, they only saw the institutions they administered in the drop down
for last institution where they should see all.

This patch adds an extra flag to the get_institution_selector to allow
non site admins to be treated as a site admin (for the purpose of getting
a list of all instituions.

Bug #825578

Change-Id: Idb0c044177dce433de00cc00338a1db25233ed6c
Signed-off-by: Hugh Davenport <hugh@catalyst.net.nz>