Remove temporary session folder code (bug #817940)
This patch removes the 'temporary' code that was checking for the
existance of the session directories at each session creation. This
check is performed at install and at update, however if the session
directory goes away somehow between updates, there is currently
_no_ way to restore it without one of these two options. This may
need to be considered before merging in this change.

Change-Id: Ia6020c0b73ce9e172e718c9638a312b1f0162295
Signed-off-by: Melissa Draper <melissa@catalyst.net.nz>