Fix whitespace issues with param_... functions
bug #838429

Basically param_... functions don't check for trailing whitespace
This patch fixes this by trimming the value after it retrieves it.

Also, the param_integer_list function says in the comment it allows
optional whitespace, but the regex expression used doesn't allow it.
This patch fixes that.

Also, the param_signed_integer function didn't check on the entire value
and you could have a value like "ab+10cd".
This patch alters the regex to capture the start and end of string ^ and $

Change-Id: Ic26cc257f0057b6b686fc8c2e3c789610e055507
Signed-off-by: Hugh Davenport <hugh@catalyst.net.nz>