Fix list of invited & requested groups on user profile
The group jointype changes from bug #610690 (which allow invitations
to any group) resulted in incorrect lists of invited/requested groups
on the user profile.  Groups the user is already a member of are
listed under 'invited to join', and controlled groups are listed under
'requested membership of'.  This change fixes these lists, and merges
the two queries that find the groups for which the logged-in user is an
admin.

Change-Id: Iff675790391ef1ada22b6e9bb5930772fd2c3bf3
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>