Enable multipleblogs when copying journals from others (bug #820182)
Currently, copying journal does not ensure that the multipleblogs
setting is set. This means that the user must manually do this to
see their copied journal entries. This patch ensures that the
setting is set on, and reloads the session. It also notifies the
user that the journal has been made.

Change-Id: Ief724d0bb7f2aaf489b995e53194afa8382819c0
Signed-off-by: Melissa Draper <melissa@catalyst.net.nz>