Reorder tabs for editing content of a page
Bug #697837

Added a database entry to blocks to enable reordering of tabs.
This sort field can be used as a weighting but must be unique.
other tabbed structures should also have a field as such to allow
admins to change the order on the fly.

Patch by Sean Brennan, cleaned and rebased by Hugh Davenport

Change-Id: I269d3f00288b5470b8e612a54527ae7f22aa747c
Signed-off-by: Hugh Davenport <hugh@catalyst.net.nz>