Trim directory names before they are made (bug #813905)
There are several places in the code where we make directories
for stuff, both in the filesystem and in the artefacts table.
If they have trailing spaces, then this can mess up users html
exports etc. This patch should make exports safe for existing
folders with whitespace at the end, and prevent new folders from
from being made with the problem.

Change-Id: Ia593f7f773e5ffe91ce74e8a736074b1fe1026b2
Signed-off-by: Melissa Draper <melissa@catalyst.net.nz>