Hide upload form when group files tab disables it (bug #854369)
When selecting files for a block, the filebrowser "owner" tabs (my
files, group files, etc.) are supposed to determine whether the upload
form is displayed.

However, when browsing group files, the tab upload rules are
overridden by code introduced in commit 3ed6f3390c to show/hide the
upload form depending on the group folder permissions.  So when
configuring a block within a personal view, the upload form should be
hidden when browsing group files, but it's not hidden when the user
has permission to edit the selected group folder.

This patch ensures the filebrowser js remembers the state of the
upload permission granted by the selected tab, and doesn't override
that status when changing to a folder with different group edit
permissions.

Change-Id: I673815e17175508fc9d90e78b9e0a25e11831935
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>