Hide form when group file permissions doesn't allow upload (bug #854377)
When configuring a file block, the upload form should be hidden
whenever the logged in user is browsing inside a group folder that
s/he cannot edit. However, this doesn't happen correctly for the first
folder you browse to when editing a group view, so if that first
folder doesn't allow editing, the upload form is displayed anyway.

This occurs because the group id is not passed through along with the
change folder command when you're editing a group view.

Change-Id: Ife148dee1ec1908b10dd90cc28da1c962bff39a0
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>