Don't allow friend suggestions for certain groups (bug #845290)
When a group is not open and doesn't allow membership requests, there
is no point in allowing members to suggest the group to their friends,
because there's nothing those friends can do about the suggestion.

The 'suggest' button already appears only for open & request groups,
but this change adds an error on the edit group page when a user tries
to set suggestfriends on a non-open, non-request group.

Change-Id: Ie8f468b9f04ca7e70570b24e1b405634aa1f1fe6
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>