Create all profiles with logged-in access by default (bug #807278)
There is currently only a site-wide setting for logged-in profile view
access.  This means that when two institutions share a site, and one
wants open profiles and another only wants institution-visible profiles,
new users in the open-profile institution must be asked to manually add
logged-in access on their profile.  Because we expect many more
open-profile institutions than closed-profile institutions on a site, it
makes more sense to put the onus on the closed-profile institutions to
ensure their members remove logged-in access manually.

This change ensures logged-in access is added to all new profiles, but
only locks that access when the loggedinprofileviewaccess setting is on.

Change-Id: I3375d7b2f8eb635a11879bf7758267f48f74c508
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>