"Display mode" on goals, skills, & interests pages (bug #869658)
Instead of displaying content inside the editors on these pages,
it is useful if the user can see the html content as it would be
displayed to someone else.  The default on these pages is now to
display the various resume fields as html, with an edit button
beside each field to open up the editor.

Also combines some code to create & submit several similar resume
forms into functions in artefact/resume/lib.php

Change-Id: I649d7788a7e70dfd5dd3eb1ce648f37a7c057aa0
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>