Fix user csv updates when the db has duplicate emails (bug #862933)
The patch to update existing users by csv (commit 2f632f7988) uses
get_record to check whether email addresses are taken, stupidly
assuming that email addresses are unique. This throws an SQLException
when there are two users already on the site who have the same email
address.

This patch allows the csv update to succeed when one of the existing
emails is already owned by the user specified in the csv line.

Change-Id: I2212907dfbd30de37cd26c6f6504989b36a574e0
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>