Allow editing of a textbox block when artefact not found (bug #736665)
When a textbox block references an artefact that has been deleted,
or has become unpublishable by the current user, clicking the block's
configure button causes an error and fails to bring up the config
form.  Avoid this by catching the ArtefactNotFound exception, and
allowing the block to be reconfigured when it contains an invalid
artefact.

Change-Id: Ia1611a6abf2328c503655dacf01b2c414a603b0a
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>