Get editable status along with artefactchooser artefacts (bug #736665)
When choosing an artefact for inclusion in a view, add a boolean
property to indicate whether each returned artefact is editable in
the context of the view.  So, individual artefacts are editable in
individual views provided the 'owner' is the same, group artefacts
are editable when the group is the same as the view's group, etc.

Change-Id: I3ec7d13d5401ddd007916f9382fd0673739dd815
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>