Return sanitised copy of description from artefactchooser (bug #736665)
When we are going to display an html artefact description outside a
tinymce window, in 'read-only' mode, we need it to be cleaned up by
the server in advance.  This patch adds returns a safe version of an
artefact description along with the original description, when
requested in the artefactchooser configuration.

Change-Id: I8725e56608214bba0a262de78e6dac8fcf5a4dd8
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>