Display textbox notes as read-only in config form (bug #736665)
When configuring a textbox that includes an html artefact owned by
someone other than the view owner, it must be made clear to the user
that the textbox cannot be edited within the current view.

In this case, the sanitised html from the artefactchooser is
displayed inside a grey box in the config form, and the tinymce editor
is hidden.

A help file has been added to explain why the user cannot edit the
selected artefact.

(When tinymce includes an toggle for its read-only mode, this can be
simplified)

Change-Id: I81b44928b7ac83ffaebba04d4ab34d3964743bf4
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>