When configuring a read-only textbox, allow copying (bug #736665)
When configuring a textbox containing a read-only artefact (or one
which has copies in other blocks), the user will often want to
create a new copy of the note instead of reusing the note.  This
patch adds a 'copy' link beside the read-only message, which will
unhide the tinymce box and set a flag to ensure that a new html
artefact gets created on submission.

Change-Id: I5214c5b0ae615cd9f32dfc5c71efc942b5961756
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>