Remove view title display logic from my views template (bug #860154)
As a result of profile view access changes in bug #807278, profile
views are displayed on the share page.  So the 'editing' version
of view titles needs to be displayed in two places (my views/
group views etc., and on the share page).  It makes sense to use
the same View methods in both cases, and remove the old logic to
get the view title and url from the template.

Change-Id: Ifa36651dc3358acd71ca63407df79799ddd47299
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>