Eval js to rewrite task titles after block configuration (bug #845534)
When a plans block is first added to the page, the javascript overlay of
the task titles was not evaluated.  This is fixed by moving javascript
functions for the plans block into a new javascript file and pulling the
script onto the page with the same method used by the configuration form
javascript. It also:

- Rewrites the task title links after the callback from getScript
- Sets the block's artefactplugin so the js file is found correctly
- Removes any onclick from task titles before rewriting (in case
  there are two plans blocks on a page)
- Hides the pagination on first load, as the paginator object is not
  present at that point.

Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>
Change-Id: I913d8a6fa1f9f2f159dc8076bbdf86bd50b4b588