Eval js to rewrite task titles after block configuration (bug #845534)
When a plans block is first added to the page, the javascript overlay of
the task titles was not evaluated.  This is fixed by moving javascript
functions for the plans block into a new javascript file and pulling the
script onto the page with the same method used by the configuration form
javascript.  A new get_instance_inline_javascript method is added to
initialise the block when it's first configured.

Change-Id: I148a1e8a667c26fe58b3f915768ada4743718c5b
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>