Fix group shortnames in share with group buttons (bug #881681)
When running through the list of a user's groups for display in the
Share with my groups buttons, a group with a long name can cause
another group's button to have the long-named group's short name
displayed.  The php array pointer is not moved when items are added
to the array.

Change-Id: If90b63ec721bfbee2b2c1ad2c7d29d855c9ce1f2
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>