Blog block pagination bug prevents images from being displayed
Bug #886581

As the viewid is not added to the URL for the images, images are
considered not published and we don't have permission to see them.
In fact, if we right click on the image place holder and click
on 'Display image' (or equivalent browser option), we are presented
with a login screen, even if we are logged in.

Just make sure the viewid is passed to ArtefactTypeBlogpost::get_posts()
too as part of the configdata, to include it in the image access URL.

Change-Id: I23138180fbd771755407c39a9669860e4ecef762
Signed-off-by: Iñaki Arenaza <iarenaza@mondragon.edu>