Add local hooks for registration and user initialisation (bug #1001064)
Three new hooks are added:

local_init_user() - called after $USER is initialised.  This is useful
for changing the user's theme before $THEME is initialised.

local_register_submit() - called when the registration form is
successfully submitted, but before the submitted values are saved to
usr_registration.  This is useful for remembering the properties or
preferences of the logged-out user when the form was submitted.

local_post_register() - called after a user has successfully been
created and logged in during registration.  This is useful when
properties of the user (which may have been saved to usr_registration
by local_register_submit()) need to be transferred to the newly
registered user.

Change-Id: Ifcb19737bdcecb550185624f2fd78e541690a337
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>