Stop blocktype images from freezing during drag in IE (bug #891916)
When dragging a blocktype in the view editor in IE, dragging it very
slowly usually causes the block to suddenly disappear, and the cursor
to turn into a crossed circle.  If you let go of the mouse, the
blocktype image appears at the cursor, and clicking the mouse again
will drop the blocktype.  This can be fixed by disabling IE's default
startdrag function.

Change-Id: I03b3150dbdef7f59f6f638d0cc71a773df5de576
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>