Ensure dragged blocktype appears above view content area in IE (bug #891916)
In IE, when dragging a blocktype image into the content area of the
view editor, the image appears to slip underneath the white box saying
"This area shows a preview of what your page will look like..." and
then out the other side again.  Setting the z-index property on the
element containing blocktype list fixes the issue.

Change-Id: I6bc28bcf790b89d7513144012e210527b8ae1710
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>