Bring back the settings sub-menu for drop-down navigation (bug #801081)
The submenu under settings in not usable with drop-down navigation,
because there's no top-level item for it to hang from.  This patch
changes it to group-style tabs within the page when drop-down
navigation is configured.  Just creating a link from the inbox to
the notification prefs is not sufficient because plugins can now
hook into the right menu (see bug #834171).

Change-Id: I7cd6009ed77d5795a9a6880e64d90f8742e47f92
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>