Remove wwwroot from notification urls (bug #849716)
Urls stored in the url field of notification_internal_activity contain
the site's wwwroot.  This leads to broken links when the wwwroot is
changed.  It also dangerously turns what should be local links into
remote ones when for example a production database is copied into a
test Mahara instance.

This patch does three things:
- Removes the wwwroot from urls passed to ActivityType constructors
- Adds the wwwroot back on the url when sending email and emaildigest
  notifications, and when displaying them on the site.
- Removes the wwwroot from the url in all existing rows of
  notification_internal_activity.

Change-Id: I00ae1abc65be79030ed4c572ce951edc43f7a034
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>