Fix display of view access rows after error on submission (bug #894592)
When you submit the view access page and there's an error in the form
submission (due to bad start/stopdates, etc.), sometimes an item is
removed from the array of access records by unsetting the array key.
Messing with the ordering of the array keys causes a javascript error
which stops the access list from being displayed.

Change-Id: Iaa686c8adecc322a6b779ee3247b28c945031854
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>