Display stopdate fields which have errors in them (bug #894592)
When you get an 'end date in the past' error from the view access page,
the offending record is not displayed because the form removes stale
records as it pulls them out of the database.  However, when we're just
validating a bad stop date entered by the user, we need to make sure
it's still displayed to avoid confusion over the error.

Change-Id: I22c8876a599a2f00280058f32b4757d8eb523ea0
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>