Fix warnings when rendering access type after error (bug #894592)
When the view access page is submitted with an error, php warnings
are emitted when the rendering the access rows.

First because the pieform element assumes that an 'accesstype' index
exists in the accesslist value, but this is only true when the
accesslist still contains database records rather than the list
submitted by the form.

Second,  disabled public access record causes a warning because
'public' isn't in the list of 'presets' allowed to be added to the
view, so we need to keep track of the allowed presets and the full
list of presets that may have been allowed in the past.

Change-Id: Ic4db20738d53a52caeb3ebef0b44660dcb7c9ee6
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>