Add notice & disable new secret urls when publicviews is off (bug #794385)
When the site or institution has disabled public views, there's a
notice on the view access page to that effect, but the user can
happily create new secret urls with no indication that they won't
work.  This patch removes the generate new secret url button, and adds
a notice to the secret url page to inform the user when they're not
allowed to create public pages.  Also adds a sentence to the site
setting to make it clearer that disabling public pages also disables
secret urls.

Change-Id: I93a36df425bf06c4748cfdbfaf62613017518872
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>