Update navigation blocks when copying collections (bug #898923)
When copying a collection, navigation blocks in the copied pages
should be updated to point at the new copy of the collection.

After all the pages are copied, all the newly copied navigation blocks
are examined to see if they navigate the collection being copied.  If
so, they are updated to navigate the new copy of the collection.  When
a navigation block points at some collection other than the one being
copied, it is left alone.

Change-Id: I4b817acfa2cbefafc37f98f3f3f7fc12e54f61c0
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>