New BrowserID auth plugin (bug #894211)
Existing users can now login through BrowserID using one of their
confirmed email addresses as long as one of the institutions they
belong to has enabled BrowserID. We are ignoring the auth
instance that these user accounts are tied to.

Automatic creation of accounts is enabled. The only outstanding
issue is to figure out what authinstance we should use for
creating accounts. However that's a general problem with the
authentication system in Mahara, it's not specific to this
plugin. Currently, we make sure that it is an institution that:

- has registrations enabled
- is not full

When an account is automatically provisioned, the username
is set to the user's email address since that's the only piece
of information we can reliably get from BrowserID.

Change-Id: I12e59271449654d2bb394511d5ce5adfb081aa11
Signed-off-by: Francois Marier <francois@catalyst.net.nz>