Add a warning when a pieform element fails to add data to <head> tag
If a pieform is built after a call to the smarty function, elements
that a pieform element wants to put into <head> (usually js files)
may be omitted and it can be hard to track down the cause of these
errors.

This patch will produce a warning (intended for the programmer) if a
pieform element that defines the _get_headdata() function is built
after the smarty function has already been called.

Change-Id: Icf120242f521d4324548ac84228fea54d75eab8c
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>