Site option to remove 'copy of ' from copied titles (bug #905067)
When pages and collections are copied, the default title of the new
copy is "Copy of <original title>". When users are copying large
collections that are intended as templates, it's annoying for them to
have to go and edit the titles of all the copies.  This change allows
the site to turn this behaviour off and just make the new title the
same as the original.

Change-Id: Ic82367ddecc6a4b9bd3c6ea5b44370ad7fe00d3e
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>