Expire users when they've been inactive for too long (bug #890929)
The "Default account inactivity time" setting allows the admin to
specify a time period after which users who have not used the site
will be unable to login, but this is not currently enforced.

This change modifies the inactivity cron job to set the expiry date to
the current date for any user who has been inactive for longer than
the 'defaultaccountinactiveexpire' period.  It also now considers the
lastaccess and ctime fields as well as the lastlogin field.

This allows the admin to reactivate inactive users by resetting their
expiry dates in account settings.

The active column on the user table is currently only used to decide
whether users should be displayed in search results, and users are set
to inactive whenever they are deleted, suspended, or expire.

Change-Id: Ieaf7a0b36865af726fc2526895146373efbb2741
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>