When open_basedir is set, don't set CURLOPT_FOLLOWLOCATION (bug #909168)
Trying to set this option fails when an open_basedir restriction is in
effect.  This patch also checks open_basedir and adds a warning to the
admin area, and in the sanity check for an unwritable dataroot, the
open_basedir directories are listed on the screen as a hint for
installation.

Change-Id: Ic394517f436955b038f321b06011b818d4967026
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>