Delete group files and views when deleting groups (bug #913536)
Most group data was not being deleted along with the group.  This
patch ensures all group views and artefacts are deleted when a group
is deleted, and that any views submitted to the group are released.
Forums are deleted using the delete method of InteractionInstance, but
this just marks the forum as deleted rather than wiping the data.

Change-Id: If160ebde3bf7c20d92929986fe70ad9e3853c6be
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>