Fix block configuration when artefactchooser has no results (bug #913571)
When the artefactchooser element returns no results in a block
configuration form, and you click on the search tab, the pagination
javascript attempts to update the result list into a missing element,
making most of the form disappear in a disconcerting manner.  This is
fixed by displaying the pagination bar even when the artefactchooser
comes up with no elements to start with.

Change-Id: I66b6660b5a3300a6b42e2d0b388175de5908f6e0
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>