Use type=profileiconbyid in more requests for thumb.php
This can help to reduce the overall number of requests for thumb.php
because the 'profileiconbyid' urls are cached for longer by the
browser.

(Where the profile_icon_url function can be used, we also avoid the
need to generate additional requests for users without profileicons,
when remote avatars are disabled.  But because profile_icon_url relies
on the nouserphoto icon being available in the requested size, we
can't use it everywhere until we either generate the nouserphoto icon
in more sizes, or change all those pages to use standard sizes.)

Change-Id: I81c5c5f4b7f9136bf9b6a6f711e55a4bceb6fa14
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>