Ensure filetype is set when creating new file artefacts (bug #914574)
When a new file artefact is created from a file that wasn't uploaded
by a browser, the filetype property is not always set.  This causes
warnings when extracting files from archives.  Moving the call to
file_mime_type right before construction of the new artefact type
fixes the problem and allows us to remove some other calls to
file_mime_type during import and upload.

Change-Id: I8e644bfc1e6bc0fbaa90f1e6637288997b0d028c
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>