Fix sql error in bulk user deletion (bug #916998)
When a user is deleted, deletion of notification_internal_activity
records can fail with a foreign key error when another row in
notification_internal_activity has a parent field pointing to one of
the records we are trying to delete.  This is fixed by setting the
parent fields to null before deletion.

The handle_event('deleteuser') call in delete_user is also moved
inside the transaction, to ensure that deleted users will have their
notifications deleted correctly.  Other subscribers to the deleteuser
event in core should be unaffected.

Change-Id: Ibc0a695e986bb175a00441f02d3e40abf38cbcde
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>