Allow staff to access the admin user search page (bug #919009)
Staff members will be able to search through all users in their
institutions, but they cannot edit a user unless they are also an
administrator for them (institution or site administrator).

Links from the usernames to the account settings page are only
displayed when the logged-in user is an administrator for the listed
user.

The "edit selected users" button is only displayed to site and
institutional admins, and the bulk actions page it links to now
explicitly states which users have been filtered out due to lack of
institutional admin privileges.

Email addresses are removed from the list for staff users.

If the logged-in user is an admin in one institution, and staff in
another, email is displayed and usernames linked only for the
'adminned' users.  The bulk actions page is still reachable, but any
selected users who are not adminned by the logged-in user will be
removed.

Change-Id: Ia65a960df9af7c5794002007e7580ba5f9a4c9d7
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>