Add admin report on access lists of user pages (bug #919009)
Adds a new "User reports" page to the admin section, accessible by
admins, which produces a list of all the pages owned by a given set of
users, and a list of who is on the access list for each page.

The users are selected using the checkboxes on the admin user search
page, and the page is reachable from admin user search using a new
"View reports" button on that page.

The CSV download that previously appeared on the Bulk actions page is
more appropriate on a reports page, so CSV download is also moved from
Bulk actions to User reports.  Email and remoteuser fields are only
displayed to site and institution administrators.

Because some sites will not want to allow staff to see the page access
lists of all users in their institutions, access to this page by staff
is controlled by a new "Staff report access" site setting.

Change-Id: Id02b58416e3dfb28fd39c1170426ddefe6669efe
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>