Allow searching and filtering by tag in view_search
Previously, the search term passed in the $query parameter would only
match in the view title and description.  Results are now returned
when this search term matches one of the view's tags.

Additionally, a new $tag parameter will act as a filter on views, so
that only views with the given tag are returned.

Change-Id: I352f31b6adb088072c8bcf45ae8b237b61f71a6f
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>