Set new profile fields when creating users from LDAP (bug #906996)
When users are automatically created from an LDAP server, the student
id & display name fields are set in the user table, but not updated as
profile fields in the artefact table, so they're not displayed in the
Edit profile page.

Passing the student id & display name from the auth plugin's
get_user_info function into create_user will ensure they're set as
profile fields as well.

Change-Id: I3ae769394d2f65577f26759309affdfd59070302
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>