Display locked email as uneditable on Edit profile page (bug #917935)
The 'emaillist' pieform element doesn't appear greyed out or
uneditable when it's disabled.  This means that when the email address
field is locked in institution settings, or because the email address
is imported from Moodle, users may still try to add new addresses
without receiving a useful error message.

When displaying this element on the Edit Profile page, the help item
is hidden whenever the element is locked, because it refers to buttons
and links that don't appear in the form.

Change-Id: Ia9a149e9110932ebac9202a1691d29799d327dc2
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>