Fix file block config edit, select, upload buttons in Chrome (bug #819102)
For filebrowsers inside configuration forms in Chrome, the edit,
select, and file upload buttons appear to have no effect when the
copyright agreement is enabled.

In fact the javascript for these buttons *is* executed, but Chrome
won't display the resulting changes, making it rather difficult to
click on some of the buttons in forms that are opened up.

Adding a slight delay before disabling the file input fixes the issue.

Change-Id: I8cfde0056f0aefc027e3a532488be3c369292b65
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>