Make clean_html more forgiving when html filters are missing
When an html filter is removed from disk for some reason, but is still
listed in the config table, you get a fatal error.  If an upgrade
removes a filter, this will result in a broken-looking page before the
upgrade is run.  We don't need to throw a fatal error when a filter is
missing, the site will run just fine with more html stripped out.

Change-Id: Ic6498f368561c2e005d44562d518bde94bd3457c
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>