Collection title overrides page title (Bug 896890)
Changes htdocs/view/view.php
If a page is in a collection with tabbed navigation, the
collection title only is displayed. If a page is in a
collection without tabbed navigation OR not in a collection
at all, only the title of the page is displayed.

The author has been removed from the page header (title
used to be "PageTitle by Author" but is now just "PageTitle")
(This includes editing a page)
It will be added back with my fix for Bug 897057.

The title is also no longer a link as without small headers
it used to act as a link to the current page.

Note: The $title variable is now global and is set only once
(rather than twice within two if statements).

Change-Id: I36fa6945862d57c8e3767828e1a6c4bd6a1e3a94
Signed-off-by: Jono Mingard <reason.koan@gmail.com>