Fix get_string calls for duplicate remote username message (bug #968697)
When uploading users from a csv file, the remote usernames given in
the file are checked for duplicates.  When a duplicate is found,
get_string is called to generate the error message using the string
key 'uploadcsverrorremoteusertaken', but in two of three instances, it
is called with too few arguments to sprintf.  Sometimes this results
in no error being reported to the user at all.

The error is fixed here by creating a new string without the second
argument for the first call, and by passing the additional argument in
the second call.

Change-Id: Iff93385dee206394c74b1d6b1ad9484e90026dd4
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>