Make download.php publicly accessible (bug #979538)
In commit 647a99fd1025ceb748f8455b82418b03d788f9a8 (see bug #914490),
the sitemaps were made available from the download.php script, but
this script is not publicly accessible, so crawlers would not be able
to download them.

Making the script public is okay here, because whenever a non-sitemap
file is requested, there is already an exception thrown if the user is
not logged in.

Change-Id: Ia9c62940ee7dada05f4f1b448ead0c146171535c
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>