Allow html to appear between label and element in pieforms
In some cases, it's useful to put explanatory text or html before a
pieform element, but after the label.  A 'prehtml' field is added to
the pieform renderers to achieve this.

This will be useful, for example, when the user is able to edit the
very last portion of a url, but not the full url.  The initial,
uneditable part can be displayed before the textbox, so that the user
is aware of the context for the field they're editing.

Change-Id: I721513ae7022401e959c5f8cddd9b45d9704a8b1
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>