Add admin page to generate clean URLs
Bug #1006634, part 1

Clean urls can be enabled at any time.  Initialisation for existing
users should not be tied to a site upgrade, because it is slow, and
most sites won't ever enable it.  Instead, add an admin page with a
button to regenerate all clean urls based on existing usernames, group
names, and page titles.

Change-Id: Id8bae56a6500b1b0038fc20aacadf4817e6e922a
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>