Change version of safe iframe sources upgrade (bug #971282)
We have just committed to master a database upgrade with a version
number less than the current version number of the 1.5 branch.  This
means that an upgrade from the current 1.5 to master will not add the
safeiframe tables.

This patch changes the version number on master to ensure the upgrade
is applied.

This occurred because the 1.5_STABLE branch's version leapfrogged over
the master version during the release candidate stage.  This will be
fixed in the release script (see bug #988682).

Change-Id: Ic2929fa9f17719a6068494ab63e7f00558c2fdcc
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>