Remove email verification in registration with approval (bug #918431)
When an institution requires approval before registration, treat the
approval email as address verification.  After filling in the
registration form with an institution that requires approval, the user
is simply asked to wait, rather than clicking on an email link to
verify their email address.  After approval, clicking on the link from
the approval email both verifies the email & creates the account.

Change-Id: I43c28510665b6f2bf9302099b6bd294a3715a85d
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>