Use get_user_for_display in View::get_owner_object()
Bug #1006634, part 3

This is more efficient on many pages, because the user object has
already been fetched, or is about to be fetched, to run display_name
or profile_icon_url.

However, user object returned by get_user_for_display won't allow
can_view_view to tell whether a user is suspended or not, so
suspendedctime is added to this object (it's less identifying than
suspendedcusr).

Change-Id: Ic9d1e5d437aedbc45f6328a05997812dcf45c8d6
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>