Add support in User for theme preference
Part of bug #793308

If a user has the 'theme' property set in their account preferences,
this is used instead of the site or institution's theme.

The LiveUser::reset_institutions() function, which recalculates the
institution theme, is now called in place of LiveUser::update_theme()
whenever a user's institutions have changed.  reset_institutions() now
calls update_theme() if the user is a LiveUser.

Change-Id: I75b36da85a5aa249c3098078b8588b8a20ac9b48
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>