Preserve registration 'extra' values during approval (bug #1001064)
When approving registration, there is an option to add institution
staff permission - this property is saved in the 'extra' column of the
usr_registration table.  But if the local registraion submit hook has
been used to store other properties in this column, they are removed
during approval.  This change ensures the existing properties are left
in place.

Change-Id: Icfa57f446e88ecee5213fb4ce6ff4fdb9f780f6a
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>